requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
groq>=0.9.0
pyyaml>=6.0.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
pandas>=2.0.0
plotly>=5.18.0
pytest>=7.0.0
